refactor: split PostViewPanel into modular components
