+++ src/shared/components/PostViewPanel/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PostViewPanel";
